Homework/hw2/hw2p2/requirements.txt
numpy==1.18.1
pillow==7.0.0
pandas==1.0.0
scikit-learn==0.22.1